Remove more clientside logic from callback page
