+++ app/[locale]/callback/page.tsx
@@ -2,9 +2,11 @@ import paytrailService from '@/utils/paytrail';
 
 import { serverFetchAPI } from '@/lib/serverApi';
 import { CallbackPageFields } from '@/utils/models';
-import { updateOrderState } from '@/app/api/createPayment/route';
+import { updateOrderState } from '@/utils/helpers';
 
-import Result from './Result';
+import CallbackResetHandler from './CallbackResetHandler';
+import { getTranslation } from '@/utils/translationHelper';
+import Link from 'next/link';
 
 type CheckoutStatus = 'new' | 'ok' | 'fail' | 'pending' | 'delayed';
 
@@ -19,8 +21,6 @@ const getContent = async (locale: string) => {
   }
 }
 
-
-
 type Props = {
   params: {
     locale: string
@@ -28,8 +28,20 @@ type Props = {
   searchParams: Record<string, string>;
 }
 
+type ErrorLinkProps = {
+  locale: string;
+  translation: Record<string,string>
+}
+const ErrorLink = ({locale, translation}: ErrorLinkProps) => 
+  <Link className='text-primary-900 dark:text-primary-100 underline'
+        href={`/${locale}/summary`}>
+        {translation.backToOrder}
+</Link>
+
+
 const CallbackPage = async ({params: {locale}, searchParams}: Props) => {
   const params = searchParams;
+  const translation = await getTranslation(locale);
   const isValid = paytrailService.verifyPayment(params);
   if(isValid) {
     await updateOrderState(
@@ -40,12 +52,21 @@ const CallbackPage = async ({params: {locale}, searchParams}: Props) => {
   }
   const paymentStatus = params['checkout-status'] as CheckoutStatus;
   const content = await getContent(locale);
+  if(!content) return <div>Error on rendering callback page</div>
   
-  return <Result
-            locale={locale}
-            isValid={isValid}
-            paymentStatus={paymentStatus}
-            content={content}/>
+  if (!isValid) return <div> 
+    {content.attributes.onError} <ErrorLink locale={locale} translation={translation}/>
+    </div>
+  else if (paymentStatus !== 'ok') return <div>
+    {content.attributes.onCancel} <ErrorLink locale={locale} translation={translation}/>
+  </div>
+  
+  return (
+    <div className='container text-primary-900 dark:text-primary-100 max-w-3xl bg-secondary-50 dark:bg-secondary-800 mx-auto rounded shadow-md p-2 pt-4 sm:p-8'>
+      <CallbackResetHandler isValid={isValid} paymentStatus={paymentStatus}/>
+      <p>{content.attributes.onSuccess}</p>
+    </div>
+  );
 }
 
 export default CallbackPage
